requirements.txt
python-multipart
fastapi[standard]
ruff
opencv-python
numpy
keras>=3.5
tensorflow
pymongo
motor
pydantic-settings
gdown
beanie
httpx
pytest
